feat(rbac): introduce silent Authorize for internal checks without audit logging
